NO-TICKET: types: move system_contract_type module into system
