+++ execution_engine_testing/tests/src/test/storage_costs.rs
@@ -28,8 +28,10 @@ use casper_types::{
 #[cfg(not(feature = "use-as-wasm"))]
 use casper_types::{
     runtime_args,
-    system::auction::{self, DelegationRate},
-    system_contract_type::AUCTION,
+    system::{
+        auction::{self, DelegationRate},
+        AUCTION,
+    },
 };
 
 use num_rational::Ratio;
