+++ types/src/lib.rs
@@ -45,7 +45,6 @@ mod protocol_version;
 pub mod runtime_args;
 mod semver;
 pub mod system;
-pub mod system_contract_type;
 mod tagged;
 mod transfer;
 mod transfer_result;
@@ -77,7 +76,6 @@ pub use phase::{Phase, PHASE_SERIALIZED_LENGTH};
 pub use protocol_version::{ProtocolVersion, VersionCheckResult};
 pub use runtime_args::{NamedArg, RuntimeArgs};
 pub use semver::{SemVer, SEM_VER_SERIALIZED_LENGTH};
-pub use system_contract_type::SystemContractType;
 pub use tagged::Tagged;
 pub use transfer::{DeployHash, Transfer, TransferAddr, DEPLOY_HASH_LENGTH, TRANSFER_ADDR_LENGTH};
 pub use transfer_result::{TransferResult, TransferredTo};
